Complete

+++ Tiki_Shop/src/main/java/com/qnit18/Tiki_Shop/repo/ImageRepository.java
@@ -3,5 +3,5 @@ package com.qnit18.Tiki_Shop.repo;
 import com.qnit18.Tiki_Shop.entity.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageRepository extends JpaRepository<Image, Integer> {
+public interface ImageRepository extends JpaRepository<Image, Long> {
 }
